Split core logic into lib.rs
