+++ apps/rusti/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rusti"
+name = "dicephrase"
 version = "0.1.0"
 edition = "2021"
